+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Autonomous/Unneeded/autoNearNetBlue.java
@@ -27,7 +27,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.firstinspires.ftc.teamcode.Autonomous;
+package org.firstinspires.ftc.teamcode.Autonomous.Unneeded;
 
 import com.qualcomm.hardware.rev.RevHubOrientationOnRobot;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
@@ -40,8 +40,6 @@ import com.qualcomm.robotcore.util.Range;
 
 import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
 import org.firstinspires.ftc.robotcore.external.navigation.YawPitchRollAngles;
-import org.firstinspires.ftc.teamcode.Assemblies.LimbArm;
-import org.firstinspires.ftc.teamcode.Assemblies.SampleClaw;
 
 /*
  *  This OpMode illustrates the concept of driving an autonomous path based on Gyro (IMU) heading and encoder counts.
@@ -91,19 +89,19 @@ import org.firstinspires.ftc.teamcode.Assemblies.SampleClaw;
  *  Remove or comment out the @Disabled line to add this OpMode to the Driver Station OpMode list
  */
 
-@Autonomous(name="ImuStrafeFix", group="Robot")
+@Autonomous(name="BlueNet", group="Robot")
 @Disabled
-public class AutonomousFixedStrafe extends LinearOpMode {
+public class autoNearNetBlue extends LinearOpMode {
 
     /* Declare OpMode members. */
     private DcMotor leftFrontDrive   = null;
     private DcMotor rightFrontDrive = null;
     private DcMotor rightBackDrive = null;
     private DcMotor leftBackDrive = null;
-    private IMU             imu         = null;
-    private ElapsedTime Time = new ElapsedTime();
-    // Control/Expansion Hub IMU
+    private IMU             imu         = null;      // Control/Expansion Hub IMU
+
     private double          headingError  = 0;
+
     // These variable are declared here (as class members) so they can be updated in various methods,
     // but still be displayed by sendTelemetry()
     private double  targetHeading = 0;
@@ -111,12 +109,10 @@ public class AutonomousFixedStrafe extends LinearOpMode {
     private double  turnSpeed     = 0;
     private double  leftSpeed     = 0;
     private double  rightSpeed    = 0;
-    private int     leftFrontTarget    = 0;
-    private int leftBackTarget   = 0;
-    private int rightBackTarget  = 0;
-    private int rightFrontTarget = 0;
-    LimbArm arm;
-    SampleClaw Claw;
+    private int     leftTarget    = 0;
+    private int     rightTarget   = 0;
+    private int     SecondleftTarget    = 0; //Only used for strafing
+    private int     SecondrightTarget   = 0; //
     // Calculate the COUNTS_PER_INCH for your specific drive train.
     // Go to your motor vendor website to determine your motor's COUNTS_PER_MOTOR_REV
     // For external drive gearing, set DRIVE_GEAR_REDUCTION as needed.
@@ -133,7 +129,7 @@ public class AutonomousFixedStrafe extends LinearOpMode {
     // They can/should be tweaked to suit the specific robot drive train.
     static final double     DRIVE_SPEED             = 0.7;     // Max driving speed for better distance accuracy.
     static final double     TURN_SPEED              = 0.4;     // Max turn speed to limit turn rate.
-    static final double     HEADING_THRESHOLD       = 2.0 ;    // How close must the heading get to the target before moving to next step.
+    static final double     HEADING_THRESHOLD       = 5.0 ;    // How close must the heading get to the target before moving to next step.
                                                                // Requiring more accuracy (a smaller number) will often make the turn take longer to get into the final position.
     // Define the Proportional control coefficient (or GAIN) for "heading control".
     // We define one value when Turning (larger errors), and the other is used when Driving straight (smaller errors).
@@ -142,70 +138,28 @@ public class AutonomousFixedStrafe extends LinearOpMode {
     static final double     P_TURN_GAIN            = 0.02;     // Larger is more responsive, but also less stable.
     static final double     P_DRIVE_GAIN           = 0.03;     // Larger is more responsive, but also less stable.
 
-    public void Orientations() {
-        leftFrontDrive.setDirection(DcMotor.Direction.REVERSE);
-        leftBackDrive.setDirection(DcMotor.Direction.REVERSE);
-        rightFrontDrive.setDirection(DcMotor.Direction.FORWARD);
-        rightBackDrive.setDirection(DcMotor.Direction.FORWARD);
-        Wait(0.1);
-    }
-
-    public void flip(String FirstLocation, String SecondLocation) {
-        // The input is expected to be L or R for the first value and F or B for the second value
-        if (FirstLocation.equals("L")) {
-            if (SecondLocation.equals("F")) {
-                if (leftFrontDrive.getDirection() == DcMotor.Direction.REVERSE) {
-                    leftFrontDrive.setDirection(DcMotor.Direction.FORWARD);
-                } else {
-                    leftFrontDrive.setDirection(DcMotor.Direction.REVERSE);
-                }
-            } else {
-                if (leftBackDrive.getDirection() == DcMotor.Direction.REVERSE) {
-                    leftBackDrive.setDirection(DcMotor.Direction.FORWARD);
-                } else {
-                    leftBackDrive.setDirection(DcMotor.Direction.REVERSE);
-                }
-            }
-        } else {
-            if (SecondLocation.equals("F")) {
-                if (rightFrontDrive.getDirection() == DcMotor.Direction.REVERSE) {
-                    rightFrontDrive.setDirection(DcMotor.Direction.FORWARD);
-                } else {
-                    rightFrontDrive.setDirection(DcMotor.Direction.REVERSE);
-                }
-            } else {
-                if (rightBackDrive.getDirection() == DcMotor.Direction.REVERSE) {
-                    rightBackDrive.setDirection(DcMotor.Direction.FORWARD);
-                } else {
-                    rightBackDrive.setDirection(DcMotor.Direction.REVERSE);
-                }
-            }
-        }
-        Wait(0.1);
-    }
-
 
     @Override
     public void runOpMode() {
 
-        arm = new LimbArm(hardwareMap, telemetry);
-        Claw = new SampleClaw(hardwareMap);
-
         // Initialize the drive system variables.
         leftFrontDrive  = hardwareMap.get(DcMotor.class, "leftFront");
         rightFrontDrive = hardwareMap.get(DcMotor.class, "rightFront");
-        rightBackDrive  = hardwareMap.get(DcMotor.class, "rightBack");
-        leftBackDrive   = hardwareMap.get(DcMotor.class, "leftBack");
+        rightBackDrive = hardwareMap.get(DcMotor.class, "rightBack");
+        leftBackDrive = hardwareMap.get(DcMotor.class, "leftBack");
         // To drive forward, most robots need the motor on one side to be reversed, because the axles point in opposite directions.
         // When run, this OpMode should start both motors driving forward. So adjust these two lines based on your first test drive.
-        // Note: The settings here assume direct drive on left and right wheels.  Gear Reduction or 90 Deg drives may require direction flip
-        Orientations();
+        // Note: The settings here assume direct drive on left and right wheels.  Gear Reduction or 90 Deg drives may require direction flips
+        leftFrontDrive.setDirection(DcMotor.Direction.REVERSE);
+        leftBackDrive.setDirection(DcMotor.Direction.REVERSE);
+        rightFrontDrive.setDirection(DcMotor.Direction.REVERSE);
+        rightBackDrive.setDirection(DcMotor.Direction.REVERSE);
         /* The next two lines define Hub orientation.
          * The Default Orientation (shown) is when a hub is mounted horizontally with the printed logo pointing UP and the USB port pointing FORWARD.
          *
          * To Do:  EDIT these two lines to match YOUR mounting configuration.
          */
-        RevHubOrientationOnRobot.LogoFacingDirection logoDirection = RevHubOrientationOnRobot.LogoFacingDirection.RIGHT;
+        RevHubOrientationOnRobot.LogoFacingDirection logoDirection = RevHubOrientationOnRobot.LogoFacingDirection.LEFT;
         RevHubOrientationOnRobot.UsbFacingDirection  usbDirection  = RevHubOrientationOnRobot.UsbFacingDirection.UP;
         RevHubOrientationOnRobot orientationOnRobot = new RevHubOrientationOnRobot(logoDirection, usbDirection);
 
@@ -223,11 +177,7 @@ public class AutonomousFixedStrafe extends LinearOpMode {
         rightFrontDrive.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
         leftFrontDrive.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
         rightBackDrive.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
-        leftFrontDrive.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-        rightFrontDrive.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-        leftBackDrive.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-        rightBackDrive.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-        imu.resetYaw();
+
         // Wait for the game to start (Display Gyro value while waiting)
         while (opModeInInit()) {
             telemetry.addData(">", "Robot Heading = %4.0f", getHeading());
@@ -235,8 +185,23 @@ public class AutonomousFixedStrafe extends LinearOpMode {
         }
 
         // Set the encoders for closed loop speed control, and reset the heading.
-        // BEGIN AUTO CODE //
+        leftFrontDrive.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+        rightFrontDrive.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+        leftBackDrive.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+        rightBackDrive.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+        imu.resetYaw();
 
+        StrafeRobot(DRIVE_SPEED, 5);
+        driveStraight(DRIVE_SPEED, 32, 0);
+        driveStraight(DRIVE_SPEED, -28, 0);
+        turnToHeading(TURN_SPEED, -90);
+        driveStraight(DRIVE_SPEED, 55, -90);
+        turnToHeading(TURN_SPEED, 180);
+        driveStraight(DRIVE_SPEED, 30, 90);
+        // Step through each leg of the path,
+        // Notes:   Reverse movement is obtained by setting a negative distance (not speed)
+        //          holdHeading() is used after turns to let the heading stabilize
+        //          Add a sleep(2000) after any step to keep the telemetry data visible for review
         telemetry.addData("Path", "Complete");
         telemetry.update();
         sleep(1000);  // Pause to display last telemetry message.
@@ -265,25 +230,25 @@ public class AutonomousFixedStrafe extends LinearOpMode {
     public void driveStraight(double maxDriveSpeed,
                               double distance,
                               double heading) {
-        Orientations();
+
         // Ensure that the OpMode is still active
         if (opModeIsActive()) {
 
             // Determine new target position, and pass to motor controller
             int moveCounts = (int)(distance * COUNTS_PER_INCH);
-            leftFrontTarget = leftFrontDrive.getCurrentPosition() + moveCounts;
-            leftBackTarget = leftBackDrive.getCurrentPosition() + moveCounts;
-            rightFrontTarget = rightFrontDrive.getCurrentPosition() + moveCounts;
-            rightBackTarget = rightBackDrive.getCurrentPosition() + moveCounts;
+            leftTarget = leftFrontDrive.getCurrentPosition() + moveCounts;
+            leftTarget = leftBackDrive.getCurrentPosition() + moveCounts;
+            rightTarget = rightFrontDrive.getCurrentPosition() + moveCounts;
+            rightTarget = rightBackDrive.getCurrentPosition() + moveCounts;
 
             // Set Target FIRST, then turn on RUN_TO_POSITION
-            // If Strafing then reverse motor directions
+           // If Strafing then reverse motor directions
 
 
-            leftFrontDrive.setTargetPosition(leftFrontTarget);
-            rightFrontDrive.setTargetPosition(rightFrontTarget);
-            leftBackDrive.setTargetPosition(leftBackTarget);
-            rightBackDrive.setTargetPosition(rightBackTarget);
+                leftFrontDrive.setTargetPosition(leftTarget);
+                rightFrontDrive.setTargetPosition(rightTarget);
+                leftBackDrive.setTargetPosition(leftTarget);
+                rightBackDrive.setTargetPosition(rightTarget);
 
 
 
@@ -299,7 +264,7 @@ public class AutonomousFixedStrafe extends LinearOpMode {
 
             // keep looping while we are still active, and BOTH motors are running.
             while (opModeIsActive() &&
-                    (leftFrontDrive.isBusy() && rightFrontDrive.isBusy() && rightBackDrive.isBusy() && leftBackDrive.isBusy())) {
+                   (leftFrontDrive.isBusy() && rightFrontDrive.isBusy() && rightBackDrive.isBusy() && leftBackDrive.isBusy())) {
 
                 // Determine required steering to keep on heading
                 turnSpeed = getSteeringCorrection(heading, P_DRIVE_GAIN);
@@ -323,83 +288,51 @@ public class AutonomousFixedStrafe extends LinearOpMode {
             rightBackDrive.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
         }
     }
+public void StrafeRobot(double maxDriveSpeed, double distance) {
+    // Determine new target position, and pass to motor controller
+    int moveCounts = (int)(Math.abs(distance) * COUNTS_PER_INCH);
+   if (distance > 0) {
+       SecondleftTarget = leftFrontDrive.getCurrentPosition() + moveCounts;
+       leftTarget = leftBackDrive.getCurrentPosition() - moveCounts;
+       SecondrightTarget = rightFrontDrive.getCurrentPosition() - moveCounts;
+       rightTarget = rightBackDrive.getCurrentPosition() + moveCounts;
+   } else {
+       SecondleftTarget = leftFrontDrive.getCurrentPosition() - moveCounts;
+       leftTarget = leftBackDrive.getCurrentPosition() + moveCounts;
+       SecondrightTarget = rightFrontDrive.getCurrentPosition() + moveCounts;
+       rightTarget = rightBackDrive.getCurrentPosition() - moveCounts;
+   }
+    // Set Target FIRST, then turn on RUN_TO_POSITION
+    // If Strafing then reverse motor directions
+    leftFrontDrive.setTargetPosition(SecondleftTarget);
+    rightFrontDrive.setTargetPosition(SecondrightTarget);
+    leftBackDrive.setTargetPosition(leftTarget);
+    rightBackDrive.setTargetPosition(rightTarget);
 
-    public void StrafeRobot(double maxDriveSpeed, double distance, double heading) {
-      Orientations();
-      if (distance > 0) {
-        flip("L", "B");
-        flip("R", "F");
-      } else {
-        flip("L", "F");
-        flip("R", "B");
-      }
+    leftFrontDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+    rightFrontDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+    leftBackDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+    rightBackDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
 
-      telemetry.addData("Encoder Count aim", (distance * COUNTS_PER_INCH));
-        if (opModeIsActive()) {
-            leftFrontDrive.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
-            leftBackDrive.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
-            rightFrontDrive.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
-            rightBackDrive.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
-            // Determine new target position, and pass to motor controller
-            // Determine new target position, and pass to motor controller
-            int moveCounts = (int)(Math.abs(distance) * COUNTS_PER_INCH);
-            leftFrontTarget = leftFrontDrive.getCurrentPosition() + moveCounts;
-            leftBackTarget = leftBackDrive.getCurrentPosition() + moveCounts;
-            rightFrontTarget = rightFrontDrive.getCurrentPosition() + moveCounts;
-            rightBackTarget = rightBackDrive.getCurrentPosition() + moveCounts;
+    maxDriveSpeed = Math.abs(maxDriveSpeed);
+    if (maxDriveSpeed > 1) {
+maxDriveSpeed = 1;
+    }
+    leftFrontDrive.setPower(maxDriveSpeed);
+    rightFrontDrive.setPower(maxDriveSpeed);
+    leftBackDrive.setPower(maxDriveSpeed);
+    rightBackDrive.setPower(maxDriveSpeed);
 
-            // Set Target FIRST, then turn on RUN_TO_POSITION
-            // If Strafing then reverse motor directions
-            leftFrontDrive.setTargetPosition(leftFrontTarget);
-            rightFrontDrive.setTargetPosition(rightFrontTarget);
-            leftBackDrive.setTargetPosition(leftBackTarget);
-            rightBackDrive.setTargetPosition(rightBackTarget);
+    while (opModeIsActive() &&
+            (leftFrontDrive.isBusy() && rightFrontDrive.isBusy() && rightBackDrive.isBusy() && leftBackDrive.isBusy())) {
+telemetry.addData("LeftTarget", leftTarget);
+telemetry.addData("SLeftTarget", SecondleftTarget);
+telemetry.addData("RightTarget", rightTarget);
+telemetry.addData("SRightTarget", SecondrightTarget);
+telemetry.update();
+    }
 
-            leftFrontDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
-            rightFrontDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
-            leftBackDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
-            rightBackDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
-
-            // Set the required driving speed  (must be positive for RUN_TO_POSITION)
-            // Start driving straight, and then enter the control loop
-            maxDriveSpeed = Math.abs(maxDriveSpeed);
-            moveRobot(maxDriveSpeed, 0);
-
-            // keep looping while we are still active, and BOTH motors are running.
-            while (opModeIsActive() &&
-                    (leftFrontDrive.isBusy() && rightFrontDrive.isBusy() && rightBackDrive.isBusy() && leftBackDrive.isBusy())) {
-
-            // Determine required steering to keep on heading
-            turnSpeed = getSteeringCorrection(heading, P_DRIVE_GAIN);
-
-            // if driving in reverse, the motor correction also needs to be reversed
-//            if (distance < 0) {
-//                turnSpeed *= -1.0;
-//            }
-            // Apply the turning correction to the current driving speed.
-            RotateRoboto(driveSpeed, turnSpeed, distance < 0);
-
-                // Display drive status for the driver.
-                // sendTelemetry(true);
-            }
-            telemetry.addData(" LF end encoder counter", leftFrontDrive.getCurrentPosition() );
-            telemetry.addData(" LB end encoder counter", leftBackDrive.getCurrentPosition() );
-            telemetry.addData(" RF end encoder counter", rightFrontDrive.getCurrentPosition() );
-            telemetry.addData(" RB end encoder counter", rightBackDrive.getCurrentPosition() );
-            telemetry.update();
-
-            // Stop all motion & Turn off RUN_TO_POSITION
-            moveRobot(0, 0);
-            leftFrontDrive.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-            rightFrontDrive.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-            leftBackDrive.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-            rightBackDrive.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-            Orientations();
-            turnToHeading(maxDriveSpeed,heading);
-          }
-
-        }
-
+}
     /**
      *  Spin on the central axis to point in a new direction.
      *  <p>
@@ -415,13 +348,12 @@ public class AutonomousFixedStrafe extends LinearOpMode {
      *              If a relative angle is required, add/subtract from current heading.
      */
     public void turnToHeading(double maxTurnSpeed, double heading) {
-       //  Orientations();
-        double TempThreshold = 1;
+
         // Run getSteeringCorrection() once to pre-calculate the current error
         getSteeringCorrection(heading, P_DRIVE_GAIN);
 
         // keep looping while we are still active, and not on heading.
-        while (opModeIsActive() && (Math.abs(headingError) > TempThreshold)) {
+        while (opModeIsActive() && (Math.abs(headingError) > HEADING_THRESHOLD)) {
 
             // Determine required steering to keep on heading
             turnSpeed = getSteeringCorrection(heading, P_TURN_GAIN);
@@ -432,9 +364,9 @@ public class AutonomousFixedStrafe extends LinearOpMode {
             // Pivot in place by applying the turning correction
             moveRobot(0, turnSpeed);
 
-//            telemetry.addData("HeadingErr", headingError);
-//            telemetry.addData("HeadinggetThresh",TempThreshold);
-//            telemetry.update();
+            telemetry.addData("HeadingErr", headingError);
+            telemetry.addData("HeadinggetThresh",HEADING_THRESHOLD);
+            telemetry.update();
         }
 
         // Stop all motion;
@@ -527,31 +459,7 @@ public class AutonomousFixedStrafe extends LinearOpMode {
         leftBackDrive.setPower(leftSpeed);
         rightBackDrive.setPower(rightSpeed);
     }
-    public void RotateRoboto(double drive, double turn, Boolean isLeft ) {
-//        driveSpeed = drive;     // save this value as a class member so it can be used by telemetry.
-//        turnSpeed  = turn;      // save this value as a class member so it can be used by telemetry.
-//        leftSpeed  = drive - turn;
-//        rightSpeed = drive + turn;
-        leftFrontDrive.setPower(drive + ((isLeft ? 1 : -1) * turn));
-        rightFrontDrive.setPower(drive + ((isLeft ? 1 : -1) * turn));
-        leftBackDrive.setPower(drive + ((isLeft ? -1 : 1) * turn));
-        rightBackDrive.setPower(drive + ((isLeft ? -1 : 1) * turn));
 
-        // Scale speeds down if either one exceeds +/- 1.0;
-//        double max = Math.max(Math.abs(leftSpeed), Math.abs(rightSpeed));
-//        if (max > 1.0)
-//        {
-//            leftSpeed /= max;
-//            rightSpeed /= max;
-//        }
-
-
-
-//        leftFrontDrive.setPower((isLeft ? -1 : 1) * leftSpeed);
-//        rightFrontDrive.setPower((isLeft ? 1 : -1) * rightSpeed);
-//        leftBackDrive.setPower((isLeft ? 1 : -1) * leftSpeed);
-//        rightBackDrive.setPower((isLeft ? -1 : 1) * rightSpeed);
-    }
     /**
      *  Display the various control parameters while driving
      *
@@ -560,9 +468,8 @@ public class AutonomousFixedStrafe extends LinearOpMode {
     private void sendTelemetry(boolean straight) {
 
         if (straight) {
-            telemetry.addData("Heading", getHeading());
             telemetry.addData("Motion", "Drive Straight");
-            telemetry.addData("Target Pos L:R",  "%7d:%7d",      leftFrontTarget, rightFrontTarget);
+            telemetry.addData("Target Pos L:R",  "%7d:%7d",      leftTarget,  rightTarget);
             telemetry.addData("Actual Pos L:R",  "%7d:%7d:%7d:%7d",      leftFrontDrive.getCurrentPosition(),
                     rightFrontDrive.getCurrentPosition(),leftBackDrive.getCurrentPosition(),rightBackDrive.getCurrentPosition());
         } else {
@@ -580,13 +487,6 @@ public class AutonomousFixedStrafe extends LinearOpMode {
     public double getHeading() {
         YawPitchRollAngles orientation = imu.getRobotYawPitchRollAngles();
         return orientation.getYaw(AngleUnit.DEGREES);
-    }
-
-    public void Wait(double seconds) {
-        Time.reset();
-        while (Time.milliseconds()  < seconds * 1000) {
-            // doesnt need anything
-        }
     }
 }
